+++ mvc/src/main/java/nextstep/mvc/controller/tobe/ControllerScanner.java
@@ -13,7 +13,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class ControllerScanner {
-    private static final Logger log = LoggerFactory.getLogger(ControllerScanner.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(ControllerScanner.class);
 
     private final Reflections reflections;
 
@@ -21,46 +21,22 @@ public class ControllerScanner {
         this.reflections = new Reflections(basePackage);
     }
 
-    public Map<HandlerKey, HandlerExecution> getControllers() {
+    public Map<Class<?>, Object> getControllers() {
         Set<Class<?>> controllerClasses = reflections.getTypesAnnotatedWith(Controller.class);
         return instantiate(controllerClasses);
     }
 
-    private Map<HandlerKey, HandlerExecution> instantiate(Set<Class<?>> classesAnnotatedWithController) {
-        Map<HandlerKey, HandlerExecution> controllers = new HashMap<>();
-
-        for (Class<?> controllerClass : classesAnnotatedWithController) {
-            List<Method> methodsWithRequestMapping = getMethodsAnnotatedWithRequestMapping(controllerClass);
-
-            for (Method method : methodsWithRequestMapping) {
-                RequestMapping annotation = method.getAnnotation(RequestMapping.class);
-                createHandlerExecution(controllers, controllerClass, method, annotation);
+    private Map<Class<?>, Object> instantiate(Set<Class<?>> controllerClasses) {
+        Map<Class<?>, Object> controllers = new HashMap<>();
+        for (Class<?> controllerClass : controllerClasses) {
+            try {
+                Object controller = controllerClass.getDeclaredConstructor().newInstance();
+                controllers.put(controllerClass, controller);
+            } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
+                LOGGER.error("there's no controller");
+                throw new IllegalStateException("there's no controller");
             }
         }
         return controllers;
-    }
-
-    private void createHandlerExecution(Map<HandlerKey, HandlerExecution> controllers, Class<?> controllerClass, Method method, RequestMapping annotation) {
-        try {
-            String requestUri = annotation.value();
-            RequestMethod[] requestMethods = annotation.method();
-
-            for (RequestMethod requestMethod : requestMethods) {
-                HandlerKey handlerKey = new HandlerKey(requestUri, requestMethod);
-                Object handler = controllerClass.getConstructor().newInstance();
-                controllers.put(handlerKey, new HandlerExecution(handler, method));
-            }
-
-        } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
-            log.error("Exception occurred while AnnotationHandlerMapping initialization" + e.getMessage());
-            throw new IllegalStateException("Handler 초기화 실패");
-        }
-    }
-
-    private List<Method> getMethodsAnnotatedWithRequestMapping(Class<?> controllerClass) {
-        Method[] declaredMethods = controllerClass.getDeclaredMethods();
-        return Arrays.stream(declaredMethods)
-                     .filter(method -> method.isAnnotationPresent(RequestMapping.class))
-                     .collect(Collectors.toList());
     }
 }
